+++ src/Matrix.cpp
@@ -2,7 +2,8 @@
  *  Matrix.c
  *  
  *
- *  Created by and copyright of J. Andres Christen on 17/04/2007.
+ *  Created by jac on 17/04/2007.
+ *  Copyright 2007 __MyCompanyName__. All rights reserved.
  *
  */